+++ src/fld.h
@@ -16,17 +16,17 @@ class FLD
     /**
      * a private vector that stores the observed counts by length
      */
-    std::vector<double> _hist;
+    std::vector<long double> _hist;
     
     /**
-     * a private double that stores the total observed counts
+     * a private long double that stores the total observed counts
      */
-    double _num_obs;
+    long double _num_obs;
     
     /**
-     * a private double that stores the sum of the observations for quick mean calculations
+     * a private long double that stores the sum of the observations for quick mean calculations
      */
-    double _sum;
+    long double _sum;
     
     /**
      * a private int that stores the minimum observed length
@@ -39,7 +39,7 @@ public:
      * @param alpha an integer that sets the initial uniform pseudo-counts
      * @param max_val an integer that sets the maximum allowable FragMap length
      */
-    FLD(double alpha, size_t max_val);
+    FLD(long double alpha, size_t max_val);
     
     /**
      * a member function that returns the maximum allowed FragMap length
@@ -51,21 +51,21 @@ public:
      * a member function that returns the mean FragMap length
      * @return mean observed FragMap length
      */ 
-    double mean() const;
+    long double mean() const;
     
     /**
      * a member function that updates the distribution based on a new FragMap observation
      * @param length an integer for the observed FragMap length
-     * @param mass a double for the mass of the observed FragMap
+     * @param mass a long double for the mass of the observed FragMap
      */
-    void add_val(size_t len, double mass);
+    void add_val(size_t len, long double mass);
     
     /**
      * a member function that returns the probability of a given FragMap length
      * @param length an integer for the FragMap length to return the probability of
      * @return probability of observing the given FragMap length
      */ 
-    double pdf(size_t len) const;
+    long double pdf(size_t len) const;
     
     /**
      * a member function that returns the probability of any length less than 
@@ -73,7 +73,7 @@ public:
      * @param length an integer for the length to return probability of a FragMap being less than or equal to
      * @return probabilty of observing a FragMap length less than or equal to the given length
      */
-    double cdf(size_t len) const;
+    long double cdf(size_t len) const;
     
     /**
      * a member function that returns the probability of the given length conditioned on the max value
@@ -81,13 +81,13 @@ public:
      * @param max an integer specifying the max nonzero probability length to renormalize the distribution
      * @return probabilty of observing the given FragMap length conditioning on the maximum value
      */
-    double npdf(size_t len, size_t max) const;
+    long double npdf(size_t len, size_t max) const;
     
     /**
      * a member function that returns the number of observed fragmaps (including pseudo-counts)
      * @return number of observed fragments
      */ 
-    double num_obs() const;
+    long double num_obs() const;
     
     void output(std::string path) const;
